backend/requirements.txt
fastapi
uvicorn
celery
redis
sqlalchemy
python-socketio
websockets
python-multipartb